* introduce `JsArray` that is the preferred sealed array instance
* introduce `JsError`
* rename `DateInstance` to `JsDate` and bring to same namespace as other primary JS types
* remove `Null` and `Undefined`, `JsValue.Null` and `JsValue.Undefined` are already there
* make `*Instance` types and suitable constructors and prototypes internal, we'll reopen with proper naming if/when needed

This paves way to some more interesting optimizations when we can easily determine that target is plain JS object instead of needing to worry about virtual dispatch.