@lahma 2022-11-11
No real performance changes expected, yet mostly jitter.

## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|dromaeo-3d-cube|25.429 ms|0.0320 ms|6708.97 KB|
| **New** |	|	| **25.758 ms (+1%)** | **0.0228 ms** | **6264.4 KB (-7%)** |
| Old |Run|dromaeo-core-eval|6.113 ms|0.0233 ms|321.58 KB|
| **New** |	|	| **5.940 ms (-3%)** | **0.0128 ms** | **321.58 KB (0%)** |
| Old |Run|dromaeo-object-array|69.449 ms|0.2199 ms|101491.97 KB|
| **New** |	|	| **67.653 ms (-3%)** | **0.2829 ms** | **101492.02 KB (0%)** |
| Old |Run|droma(...)egexp [21]|225.945 ms|1.2071 ms|178344.74 KB|
| **New** |	|	| **248.098 ms (+10%)** | **2.1092 ms** | **172335.11 KB (-3%)** |
| Old |Run|droma(...)tring [21]|413.538 ms|32.2692 ms|1326749.24 KB|
| **New** |	|	| **417.133 ms (+1%)** | **32.0312 ms** | **1325403.01 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|69.225 ms|0.3603 ms|7658.52 KB|
| **New** |	|	| **68.395 ms (-1%)** | **0.1599 ms** | **7659.2 KB (0%)** |
| Old |Run|dromaeo-3d-cube|24.519 ms|0.1731 ms|6414.59 KB|
| **New** |	|	| **25.356 ms (+3%)** | **0.0299 ms** | **5970.03 KB (-7%)** |
| Old |Run|dromaeo-core-eval|6.298 ms|0.0338 ms|308.97 KB|
| **New** |	|	| **6.351 ms (+1%)** | **0.0154 ms** | **308.97 KB (0%)** |
| Old |Run|dromaeo-object-array|65.939 ms|0.1242 ms|101453.76 KB|
| **New** |	|	| **70.084 ms (+6%)** | **0.4451 ms** | **101454.02 KB (0%)** |
| Old |Run|droma(...)egexp [21]|240.232 ms|0.8995 ms|177918.87 KB|
| **New** |	|	| **226.961 ms (-6%)** | **2.1518 ms** | **175399.23 KB (-1%)** |
| Old |Run|droma(...)tring [21]|422.583 ms|35.7367 ms|1326513.58 KB|
| **New** |	|	| **361.259 ms (-15%)** | **16.5024 ms** | **1325211.91 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|67.557 ms|0.1466 ms|7570.14 KB|
| **New** |	|	| **72.488 ms (+7%)** | **0.2912 ms** | **7569.18 KB (0%)** |